refactor(money controls): move currency helper methods to root-level utils folder
